+++ src/database/mod.rs
@@ -1,5 +1,6 @@
 //! Read/write access to the Markdown files making up the database.
 
+mod directory;
 mod doc_links;
 pub(crate) mod document;
 mod footnotes;
@@ -9,11 +10,12 @@ mod resource;
 pub(crate) mod section;
 mod tikibase;
 
-pub(crate) use crate::database::tikibase::{FileType, Tikibase};
+pub(crate) use crate::database::tikibase::Tikibase;
+pub(crate) use directory::{Directory, FileType};
 pub(crate) use doc_links::DocLinks;
-use document::Document;
+pub(crate) use document::Document;
 pub(crate) use footnotes::{Footnote, Footnotes};
 pub(crate) use line::Line;
 pub(crate) use reference::Reference;
-use resource::Resource;
+pub(crate) use resource::Resource;
 pub(crate) use section::Section;
