+++ src/scan/image_orphaned.rs
@@ -2,7 +2,7 @@ use crate::{Issue, Location, Tikibase};
 
 pub(crate) fn scan(base: &Tikibase, resource_links: &[String]) -> Vec<Issue> {
     let mut result = Vec::new();
-    for resource in &base.resources {
+    for resource in &base.dir.resources {
         let path = resource.path.to_string_lossy();
         if !resource_links.iter().any(|rl| rl == &path) {
             result.push(Issue::OrphanedResource {
